+++ esp-hal/src/lib.rs
@@ -166,7 +166,7 @@ pub use self::soc::lp_core;
 pub use self::soc::peripherals;
 pub(crate) use self::soc::peripherals::pac;
 #[instability::unstable]
-#[cfg(any(feature = "quad-psram", feature = "octal-psram"))]
+#[cfg(feature = "psram")]
 pub use self::soc::psram;
 #[cfg(ulp_riscv_core)]
 #[instability::unstable]
@@ -547,7 +547,7 @@ pub struct Config {
     /// PSRAM configuration.
     #[cfg(any(doc, feature = "unstable"))]
     #[cfg_attr(docsrs, doc(cfg(feature = "unstable")))]
-    #[cfg(any(feature = "quad-psram", feature = "octal-psram"))]
+    #[cfg(feature = "psram")]
     pub psram: psram::PsramConfig,
 }
 
@@ -629,7 +629,7 @@ pub fn init(config: Config) -> Peripherals {
 
     crate::gpio::bind_default_interrupt_handler();
 
-    #[cfg(any(feature = "quad-psram", feature = "octal-psram"))]
+    #[cfg(feature = "psram")]
     crate::psram::init_psram(config.psram);
 
     peripherals