+++ qa-test/src/bin/psram.rs
@@ -1,9 +1,12 @@
 //! This shows how to use PSRAM as heap-memory via esp-alloc
 //!
 //! You need an ESP32, ESP32-S2 or ESP32-S3 with at least 2 MB of PSRAM memory.
+//!
+//! On ESP32-S3 you might want to set `ESP_HAL_CONFIG_PSRAM_MODE` to `octal` if
+//! the device comes with octal-SPIRAM
 
 //% CHIPS: esp32 esp32s2 esp32s3
-//% FEATURES: esp-hal/quad-psram esp-alloc/internal-heap-stats
+//% FEATURES: esp-hal/psram esp-alloc/internal-heap-stats esp-hal/log
 
 #![no_std]
 #![no_main]